Use zip_cmd() and unzip_cmd() in update_node.py as well.
